specasm,salink: Fix ld (ix+1), =exp
This syntax was understood by the parser but not by the formatter or
the linker, and so resulted in corrupted text in specasm and
incorrectly encoded instructions.  All fixed now.

To get this all to fit we had to free a bit of space in line_dump.c.
We did this by factorising an array of register names.  This saved
more memory than expected so maybe we should move all those arrays
outside of the functions in which they are used.

+++ tests/test_zx81/test.sh
@@ -205,6 +205,22 @@ if [ "$str" != "82 40" ]; then
     exit 1
 fi
 
+offset=245
+str=`od -An -j$offset -t x1 -N4 zx81 | xargs`
+if [ "$str" != "dd 36 01 41" ]; then
+    echo "Expected dd 36 01 41 at offset $offset"
+    echo "got      $str"
+    exit 1
+fi
+
+offset=249
+str=`od -An -j$offset -t x1 -N4 zx81 | xargs`
+if [ "$str" != "fd 36 01 41" ]; then
+    echo "Expected fd 36 01 41 at offset $offset"
+    echo "got      $str"
+    exit 1
+fi
+
 rm zx81
 rm *.x
